Move hardware wallet state out of navigators
