+++ src/features/hardware-wallet/navigation/constants.ts
@@ -1,0 +1,1 @@
+export const HARDWARE_WALLET_TX_NAVIGATOR_SHEET_HEIGHT = 534;
