+++ src/features/hardware-wallet/state/pairHardwareWalletState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const ledgerImportDeviceIdAtom = atom({
+  default: '',
+  key: 'ledgerImportDeviceId',
+});
